sui_indexer_alt_e2e_tests
Module coin_registry

Structs

LegacyCoinOutputs

Output from querying owned objects related to creating a new legacy currency.

Constants

DEFAULT_GAS_BUDGET

5 SUI gas budget

Functions

burn_from_currency

Burn amount from a coin using the Currency object.

burn_from_treasury

Burn amount from a coin using a TreasuryCap wrapped in a shared Treasury object.

create_dynamic_currency

Create a new dynamic currency by calling the coin’s module’s new_currency function (This is specifically designed to work with the dynamic test package).

delete_migrated_legacy_metadata

Delete the migrated legacy metadata.

finalize

Finalize a Currency<T> that was sent to the CoinRegistry during a package publish.

hide_treasury_cap

Hide the treasury cap in the Treasury object (using dynamic object field).

migrate

Migrate a legacy CoinMetadata object to the coin registry.

migrate_deny_cap

Migrate the regulated state of the currency based on its DenyCapV2.

migrate_regulated_metadata

Migrate the regulated state of the currency based on its RegulatedCoinMetadata.

publish

Publish packages/coin_registry/<package> to cluster with a fresh, funded account.

show_treasury_cap

Show the treasury cap in the Treasury object (move from dynamic object field to Option field).