sui_indexer_alt_e2e_tests::coin_registry

Function create_dynamic_currency

pub async fn create_dynamic_currency(
    cluster: &mut FullCluster,
    sender: SuiAddress,
    kp: &Ed25519KeyPair,
    coin_type: StructTag,
    gas: ObjectRef,
) -> TransactionEffects

Create a new dynamic currency by calling the coin’s module’s new_currency function (This is specifically designed to work with the dynamic test package).

cluster - The test cluster to execute the transaction on

sender and kp - The account that will sign and pay for the transaction

coin_type - The full struct tag for the coin type (contains package, module, and type name)

gas - The gas object to use for the transaction

Returns the effects of running the new_currency transaction.