sui_indexer_alt_e2e_tests::coin_registry

Function migrate

pub async fn migrate(
    cluster: &mut FullCluster,
    sender: SuiAddress,
    kp: &Ed25519KeyPair,
    outputs: &LegacyCoinOutputs,
    gas: ObjectRef,
) -> TransactionEffects

Migrate a legacy CoinMetadata object to the coin registry.

sender and kp describe the account that will sign and pay for the transaction.

outputs describes owned outputs of currency creation.

gas is the gas object to use for the transaction, which must be owned by sender.

Returns the effects of running the migration transaction.