sui_indexer_alt_e2e_tests

Module find

Functions

address_mutated

Returns the reference for the first address-owned object mutated in the effects that is not a gas payment, or an error if there is none.

address_owned

Returns the reference for the first address-owned object created in the effects, or an error if there is none.

address_owned_by

Returns the reference for the first address-owned object created in the effects owned by owner, or an error if there is none.

immutable

Returns the reference for the first immutable object created in the effects, or an error if there is none.

shared

Returns the reference for the first shared object created in the effects, or an error if there is none.