sui_indexer_alt_e2e_tests
Function find_shared

pub fn find_shared(fx: &TransactionEffects) -> Result<ObjectRef>

Returns the reference for the first shared object created in the effects, or an error if there is none.